Refactor X509 cert generation into SRP helpers
