+++ crates/uselesskey-x509/src/srp/mod.rs
@@ -1,5 +1,6 @@
 //! Single-responsibility internals for X.509 fixture generation.
 
+pub(crate) mod cert;
 pub mod chain_negative;
 pub mod derive;
 pub mod negative;
